Bootstrap material design based CSS
Playing with CSS frameworks so we can get something pretty fast.

I was hoping that using these frameworks would give us good old browser
support, seems that everything (Bootstrap, jQuery, all that trash)
only support IE10+, so this only goes that far. If we want to go further
back we are going to have to do _a lot_ more tedious frontend work.

Given that even IE10 has been dead for two and a bit years I don't think
we should go back further than that. IE9 perhaps if they throw a ton of
money at us and give us extra time to do it, and then they only expect
it to technically work there, not to look correct.

Another thing in favour of not supporting IE8/9 is that if we target
those it's going to make it hard to support mobile and tablet well. If
we use modern frontend practices and make a "progressive web app" we can
create one UI that works in all devices, both online and offline, and
works like an app on tablet. If we do this we can get a tablet "app"
without having to go to the trouble of developing a native iOS app and a
native Android app.